Move factory interfaces inside SqlTestFactory
